In experiment 1 - Image size 32 and batch size 32
In Experiment 1, there are total 6 experiments from exp 1.1 to 1.6. The objective of experiment one is to choose the most suitable image size and batch size for real waste classification model. 
Experiment 1.1 was tested with image size 32 x 32 x 3 with batch size of 32.
Dropout of 0.2 was added after maxpooling layers compared to experiment 1.0