Experiment 1.4 - Image size 64 and batch size 64
Experiment 1.4 was part of a series of experiment conducted with different image size and batch size. Experiment 1.4 was tested with image size 64 and batch size 64. The model training time was 3.02 minutes for 100 epochs. Found to be most suitable for models with more hidden layer. Image size 64 was also found to be most suitable for desk top computer without any GPU. Training the model with image size 128 takes 4 time longer, as in the case of real waste classification model, took about 12 minutes.