Experiment 4.5 – L1L2 Regularization
Experiment 4 was part of a series of experiment conducted on L1 and L2 Regularization.
L1 and L2 regularization prevent overfitting by adding a penalty term to the loss function during the training phase. In Experiment 4.5, both L1 and L2 regularization were used, L1 penalty was set 0.001 and L2 was set from 0.01 to 0.05.